+++ app/playlist_ui/trackinfodialog.h
@@ -5,7 +5,9 @@
 #include "playlist/playlist.h"
 
 #include <QDialog>
+#include <QPixmap>
 #include <QStandardItemModel>
+#include <QTableView>
 #include <memory>
 
 namespace Ui {
@@ -26,11 +28,12 @@ public:
 signals:
   void tagEditorOpened(TagEditorDialog *editor);
 
+protected:
+  bool eventFilter(QObject *watched, QEvent *event) override;
+
 private slots:
-  void on_copy(const QPoint &pos);
   void on_toolButtonOpenFileManager_clicked();
   void on_toolButtonEditTags_clicked();
-  void on_search(const QPoint &pos);
   void on_labelCoverArt_customContextMenuRequested(const QPoint &pos);
   void refresh_track(const QList<quint64> &uids);
 
@@ -38,21 +41,36 @@ private:
   Ui::TrackInfoDialog *ui;
 
   QStandardItemModel model;
+  QStandardItemModel model_tags;
+  QStandardItemModel model_file;
+  QStandardItemModel model_other;
+  QPixmap cover_art;
   Track _track;
   std::shared_ptr<Playlist::Playlist> _playlist;
   QString base_title;
   QString track_path;
   QString cover_art_path;
 
-  void setup_table(const Track &track);
-  void setup_context_menu();
-  void setup_cover_art(const Track &track);
-  void setup_lyrics(const Track &track);
+  void setup_table();
+  void setup_view(QTableView *view, QStandardItemModel *m);
+  void setup_context_menu(QTableView *view);
+  void setup_cover_art();
+  void setup_lyrics();
 
-  void add_table_row(const QString& title, const QString &content);
+  void add_general_rows();
+  void add_tags_rows();
+  void add_file_rows();
+  void add_other_rows();
 
-  QString fetch_embedded_lyrics(const Track &track) const;
-  QString fetch_sidecar_lyrics(const Track &track) const;
+  void add_table_row(QStandardItemModel &m, const QString &title, const QString &content);
+
+  void rescale_cover_art();
+
+  QString yes_no(bool v) const;
+  QString value_at(const QTableView *view, const QPoint &pos) const;
+
+  QString fetch_embedded_lyrics() const;
+  QString fetch_sidecar_lyrics() const;
   void render_lyrics(const QString &source, const QString &raw);
   void render_lyrics_state(const QString &message);
 };